Set SWITCH_DISTANCE to 144 and CENTER_TO_SWITCH to 63
The switch is actually 4 feet long, and the center of the switch is 14 feet away from the alliance station, so the switch distance should be 12 feet/144 inches (whoops, my bad). 
Each side of the switch is 3 feet wide, and they’re 6 feet apart. 4.5 feet is the distance from the middle of the whole switch to the center of either side. 6 feet is the distance from the middle to the edge of the switch. Because of physics, it’s better to place the cube closer to edge of the switch, so I picked a distance in between the center of the side and the edge, which is 5.25 feet or 63 inches.