Fix functionality of drive stick buttons (committed by Shiloh)
ReverseDrive:
Delete because it interfered with the joystick code
Does not actually reverse driving

OI:
Delete stickBtnOne in case the construction of two triggers was bad
Remove the ReverseDrive command and its import

BetterJoystick:
I tried to construct actual buttons instead of getting the raw button based on the port. I don't know if this is actually better, but it works.

GoodJoystick:
Reverses the y-value of the joystick if triggered is held