Implemented a toggle for exhange vs. switch height
The button for toggling the Elevator height after intake is on Daniel's joystick on button 5.
Without pressing the button, everything should (untested) work as it has before. Once the button is pressed, it should modify the height of which the cube is raised. It is unclear how much this will impact auto (if at all).