Adds New Autos For Faster Testing
Auto 1:
"Center to Switch 2 Cube Avoid": This is a 2 cube switch auto, which on left side, uses a new path made to dramatically avoid the stack of cubes

Auto 2:
"2 Cube Switch, Start at Switch":
This is an auto where the robot starts at the end of the first path, either on the right or left side, as if it had just dropped off a cube.

Auto 3:
"Center to Switch 1 Cube Avoid Left Side":
This auto is a 1 cube switch auto using the new path to avoid the stack of cubes.